# docs/swagger.yaml
openapi: "3.0.0"
info:
  title: SCM DEMO
  description: This Documentation will help for Developer and Tester
  version: "1.0.0"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /api/v1/user/login:
   post:
      summary: "User login"
      description: "Endpoint for user login. Returns a JWT token upon successful authentication."
      tags:
        - User Managment Feature
      requestBody:
        required: true
        content:   
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  format: username
                  description: "The user's username"
                  example: username
                password:
                  type: string
                  format: password
                  description: "The user's password"
                  example: 12345
      responses:
        200:
          description: "Login successful"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: "User successfully log in."
                  data:
                    type: object
                    properties:
                      user:
                        type: object
                        properties:
                          _id:
                            type: string
                            example: 66a92c7bfb3256a5e5a4be99
                          username:
                            type: string
                            format: username
                            example: username
                          role:
                            type: string
                            example: user
                      token:
                        type: string
                        description: "JWT token for authenticated requests"
                        example: "token"

        400:
          description: "Incorrect username or password"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    example: "Please provide your account password"
...
  /api/v1/user/test-protect:
    get:
      summary: "Test user authentication"
      description: "Checks if the current user is authenticated as staff. Returns user info if authenticated."
      tags:
        - User Managment Feature
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated as staff
        '401':
          description: Unauthorized access
  /api/v1/user/test-owner:
    get:
      summary: "Test owner authentication"
      description: "Checks if the current user is authenticated as owner. Returns user info if authenticated."
      tags:
        - User Managment Feature
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated as owner
        '401':
          description: Unauthorized access
  /api/v1/user/test-staff:
    get:
      summary: "Test staff authentication"
      description: "Checks if the current user is authenticated as staff or owner. Returns user info if authenticated."
      tags:
        - User Managment Feature
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated as staff or owner
        '401':
          description: Unauthorized access
  /api/v1/user/seller:
    get:
      summary: "Get all seller usernames"
      description: "Retrieves a list of all seller usernames."
      tags:
        - User Managment Feature
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Successfully retrieved usernames."
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: "User names have been retrived successfully"
                  data:
                    type: object
                    properties:
                      userData:
                        type: array
                        items:
                          type: object
                          properties:
                            username:
                              type: string
                              example: alice
        401:
          description: "Unauthorized access"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 401
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    example: "Invalid token.Please login again"
  /api/v1/user:
    post:
      summary: "Verify user password"
      description: "Verifies the password of the currently authenticated user. Returns a success message if verified."
      tags:
        - User Managment Feature
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - verifyPassword
              properties:
                verifyPassword:
                  type: string
                  description: "The password to verify."
                  example: password123
      responses:
        200:
          description: "Password verified successfully."
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: "Old password verified successfully. You can now update your password."
        404:
          description: "User not found."
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 404
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    example: "User not found."
        401:
          description: "Password Incorrect."
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 401
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    example: "Password Incorrect."
        400:
          description: "Please provide your account password."
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    example: "Please provide your account password."
    get:
      summary: "Get all user data"
      description: "Retrieves data for all users. Only accessible by owner."
      tags:
        - User Managment Feature
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Successfully retrieved user data."
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: "User Data have been retrived successfully."
                  data:
                    type: object
                    properties:
                      userData:
                        type: array
                        items:
                          type: object
                          properties:
                            username:
                              type: string
                              example: johndoe
        401:
          description: "Unauthorized access"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 401
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    example: "Unauthorized access."
  /api/v1/user/{id}:
    get:
      summary: "Get user data by ID"
      description: "Retrieves data for a specific user by their ID. Only accessible by owner."
      tags:
        - User Managment Feature
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The user ID
      responses:
        200:
          description: "Successfully retrieved user data."
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: "User Data retrived successfully"
                  data:
                    type: object
                    properties:
                      userData:
                        type: object
                        properties:
                          username:
                            type: string
                            example: johndoe
                          role:
                            type: string
                            example: staff
                          _id:
                            type: string
                            example: 66a7decef4c583c40145e503
        400:
          description: "Invalid ID or user not found."
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    example: "There is no user with the given ID."
        401:
          description: "Unauthorized access."
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 401
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    example: "Unauthorized access."
  /api/v1/user/owner:
    patch:
      summary: "Update owner password"
      description: "Allows the owner to update their password after verifying the old password. Requires password verification."
      tags:
        - User Managment Feature
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - newPassword
                - confirmPassword
              properties:
                newPassword:
                  type: string
                  description: "The new password for the account."
                  example: newpassword123
                confirmPassword:
                  type: string
                  description: "Confirmation of the new password."
                  example: newpassword123
      responses:
        200:
          description: "Your account password updated successfully."
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: "Your account password updated successfully."
        400:
          description: "Please provide new password and confirm password."
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    example: "Please provide new password and confirm password."
        403:
          description: "Password verification required."
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 403
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    example: "Password verification required."
  /api/v1/user/staff/{id}:
    patch:
      summary: "Update staff password"
      description: "Allows the owner to update a staff member's password."
      tags:
        - User Managment Feature
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The staff user ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - newPassword
                - confirmPassword
              properties:
                newPassword:
                  type: string
                  description: "The new password for the staff member."
                  example: newstaffpassword123
                confirmPassword:
                  type: string
                  description: "Confirmation of the new password."
                  example: newstaffpassword123
      responses:
        200:
          description: "Staff account password updated successfully."
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: "Staff account password updated successfully."
        400:
          description: "Please provide new password and confirm password."
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 400
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    example: "Please provide new password and confirm password."
        404:
          description: "User not found."
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 404
                  status:
                    type: string
                    example: fail
                  message:
                    type: string
                    example: "User not found."
##################################################
# Sale Report Feature
  # /api/v1/report/sales:
...